src/requirements.txt
panda3d
panda3d-simplepbr
panda3d-gltf
numpy
scipy